Moved targeting logic into placeTemplate option
